Внедрение генерации персонажей в клиент
