+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 53ca91cbcbf018944b56c20b1e7de3a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
