+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation/start-armors.json
@@ -6,8 +6,20 @@
       "Weight": 10
     },
 	{
-      "SchemeSid": null,
+      "SchemeSid": "scholar-robe",
       "Weight": 10
+    },
+	{
+      "SchemeSid": "leather-jacket",
+      "Weight": 10
+    },
+	{
+      "SchemeSid": "gold-armor",
+      "Weight": 2
+    },
+	{
+      "SchemeSid": null,
+      "Weight": 50
     }
   ]
 }
