+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation/start-armors.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 99b6fdfcb072941439dd36dfd157f0e2
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
