+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation/start-heads.json
@@ -6,8 +6,24 @@
       "Weight": 10
     },
 	{
-      "SchemeSid": null,
+      "SchemeSid": "leather-helmet",
       "Weight": 10
+    },
+	{
+      "SchemeSid": "wizard-hat",
+      "Weight": 10
+    },
+	{
+      "SchemeSid": "elder-hat",
+      "Weight": 3
+    },
+	{
+      "SchemeSid": "highlander-helmet",
+      "Weight": 2
+    },
+	{
+      "SchemeSid": null,
+      "Weight": 50
     }
   ]
 }
