+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation/start-heads.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8f600a0db6b2c2f40b6fee9e215ee22e
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
