+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation/start-main-weapons.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e0e99daaf90790c478a32da0c95c6815
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
