+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation/start-off-weapons.json
@@ -10,6 +10,10 @@
       "Weight": 10
     },
 	{
+      "SchemeSid": "mace",
+      "Weight": 10
+    },
+	{
       "SchemeSid": "wooden-shield",
       "Weight": 10
     },
