+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation/start-off-weapons.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5a876313fa351b5478b9e86273fb1404
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
