+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation/start-props.json
@@ -20,8 +20,20 @@
       "Weight": 10
     },
 	{
-      "SchemeSid": null,
+      "SchemeSid": "battle-axe",
       "Weight": 10
+    },
+	{
+      "SchemeSid": "short-sword",
+      "Weight": 10
+    },
+	{
+      "SchemeSid": "mace",
+      "Weight": 10
+    },
+	{
+      "SchemeSid": null,
+      "Weight": 100
     }
   ]
 }
