+++ Zilon.Client/Assets/Resources/Schemes/DropTables/person-creation/start-props.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8ddf6a4e33305314082aa88b47163ec2
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
